Fix : #32 `CommandAcceptanceException` 예외 발생 수정

+++ src/main/java/com/toy/sakila/category/domain/Category.java
@@ -17,6 +17,6 @@ public class Category {
     @Value
     @AllArgsConstructor(staticName = "of")
     public static class CategoryId {
-        Long value;
+        Byte value;
     }
 }
