+++ src/main/java/com/toy/sakila/city/application/port/in/CityUpdateCommand.java
@@ -9,5 +9,5 @@ import lombok.*;
 @EqualsAndHashCode(callSuper = false)
 public class CityUpdateCommand {
     String city;
-    Integer countryId;
+    Short countryId;
 }
